clean up cart handling

+++ CoreLib/Assembler.cpp
@@ -189,13 +189,7 @@ void Assembler::disassemble_one(const uint8_t* code, Dumper& dump) {
 
 //-----------------------------------------------------------------------------
 
-void Assembler::disassemble(
-    const uint8_t* code, size_t code_size,
-    uint16_t code_base,
-    uint16_t pc,
-    int opcount,
-    Dumper& dump,
-    bool /*collapse_nops*/) {
+void Assembler::disassemble(const uint8_t* code, int code_size, int code_base, int pc, int opcount, Dumper& dump, bool /*collapse_nops*/) {
 
   size_t code_cursor = pc - code_base;
 
@@ -273,7 +267,7 @@ void Assembler::disassemble(Dumper& dump) {
     dump(".org $%04x\n", addr);
 
     const blob& code = block.second;
-    disassemble(code.data(), code.size(), addr, addr, 100000, dump);
+    disassemble(code.data(), (int)code.size(), addr, addr, 100000, dump);
   }
 }
